+++ integration-tests/api/jest.config.js
@@ -1,7 +1,10 @@
 // API
+process.chdir(__dirname)
 
 module.exports = {
+  name: "api",
   testEnvironment: `node`,
+  rootDir: "./",
   testPathIgnorePatterns: [
     `/examples/`,
     `/www/`,
@@ -11,6 +14,7 @@ module.exports = {
     `__testfixtures__`,
     `.cache`,
   ],
+  transformIgnorePatterns: [`/dist`],
   transform: { "^.+\\.[jt]s$": `../../jest-transformer.js` },
   setupFilesAfterEnv: ["../setup.js"],
   globalSetup: "../globalSetup.js",