+++ packages/medusa-payment-paypal/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/medusajs/medusa/compare/medusa-payment-paypal@1.0.40...medusa-payment-paypal@1.1.1) (2022-02-28)
+
+**Note:** Version bump only for package medusa-payment-paypal
+
+
+
+
+
+# [1.1.0](https://github.com/medusajs/medusa/compare/medusa-payment-paypal@1.0.40...medusa-payment-paypal@1.1.0) (2022-02-25)
+
+**Note:** Version bump only for package medusa-payment-paypal
+
+
+
+
+
 ## [1.0.40](https://github.com/medusajs/medusa/compare/medusa-payment-paypal@1.0.39...medusa-payment-paypal@1.0.40) (2022-02-06)
 
 ### Bug Fixes