+++ packages/medusa-plugin-algolia/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/medusajs/medusa/compare/medusa-plugin-algolia@0.0.8...medusa-plugin-algolia@0.1.1) (2022-02-28)
+
+**Note:** Version bump only for package medusa-plugin-algolia
+
+
+
+
+
+# [0.1.0](https://github.com/medusajs/medusa/compare/medusa-plugin-algolia@0.0.8...medusa-plugin-algolia@0.1.0) (2022-02-25)
+
+**Note:** Version bump only for package medusa-plugin-algolia
+
+
+
+
+
 ## [0.0.8](https://github.com/medusajs/medusa/compare/medusa-plugin-algolia@0.0.7...medusa-plugin-algolia@0.0.8) (2022-02-06)
 
 ### Bug Fixes