+++ packages/medusa-plugin-sendgrid/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/medusajs/medusa/compare/medusa-plugin-sendgrid@1.1.38...medusa-plugin-sendgrid@1.2.1) (2022-02-28)
+
+
+### Bug Fixes
+
+* reset user password subscription in sendgrid plugin ([#1072](https://github.com/medusajs/medusa/issues/1072)) ([91263fb](https://github.com/medusajs/medusa/commit/91263fbc8e70a8566a496cce97ed6256e559a786))
+
+
+### Features
+
+* new tax api ([#979](https://github.com/medusajs/medusa/issues/979)) ([47588e7](https://github.com/medusajs/medusa/commit/47588e7a8d3b2ae2fed0c1e87fdf1ee2db6bcdc2)), closes [#885](https://github.com/medusajs/medusa/issues/885) [#896](https://github.com/medusajs/medusa/issues/896) [#911](https://github.com/medusajs/medusa/issues/911) [#945](https://github.com/medusajs/medusa/issues/945) [#950](https://github.com/medusajs/medusa/issues/950) [#951](https://github.com/medusajs/medusa/issues/951) [#954](https://github.com/medusajs/medusa/issues/954) [#969](https://github.com/medusajs/medusa/issues/969) [#998](https://github.com/medusajs/medusa/issues/998) [#1017](https://github.com/medusajs/medusa/issues/1017) [#1110](https://github.com/medusajs/medusa/issues/1110)
+
+
+
+
+
+# [1.2.0](https://github.com/medusajs/medusa/compare/medusa-plugin-sendgrid@1.1.38...medusa-plugin-sendgrid@1.2.0) (2022-02-25)
+
+
+### Bug Fixes
+
+* reset user password subscription in sendgrid plugin ([#1072](https://github.com/medusajs/medusa/issues/1072)) ([90121dd](https://github.com/medusajs/medusa/commit/90121dd19d5419e239ea5d1b6feda9e7bc754d63))
+
+
+### Features
+
+* new tax api ([#979](https://github.com/medusajs/medusa/issues/979)) ([c56660f](https://github.com/medusajs/medusa/commit/c56660fca9921a3f3637bc137d9794781c5b090f)), closes [#885](https://github.com/medusajs/medusa/issues/885) [#896](https://github.com/medusajs/medusa/issues/896) [#911](https://github.com/medusajs/medusa/issues/911) [#945](https://github.com/medusajs/medusa/issues/945) [#950](https://github.com/medusajs/medusa/issues/950) [#951](https://github.com/medusajs/medusa/issues/951) [#954](https://github.com/medusajs/medusa/issues/954) [#969](https://github.com/medusajs/medusa/issues/969) [#998](https://github.com/medusajs/medusa/issues/998) [#1017](https://github.com/medusajs/medusa/issues/1017) [#1110](https://github.com/medusajs/medusa/issues/1110)
+
+
+
+
+
 ## [1.1.38](https://github.com/medusajs/medusa/compare/medusa-plugin-sendgrid@1.1.37...medusa-plugin-sendgrid@1.1.38) (2022-01-11)
 
 **Note:** Version bump only for package medusa-plugin-sendgrid