+++ packages/medusa-react/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/medusajs/medusa/compare/medusa-react@0.1.5...medusa-react@0.2.1) (2022-02-28)
+
+
+### Bug Fixes
+
+* use /admin/returns/:id/receive for swap returns ([#1041](https://github.com/medusajs/medusa/issues/1041)) ([7a3a183](https://github.com/medusajs/medusa/commit/7a3a1837a1db067c3629f1dfd7c6a95a56d649ca))
+
+
+### Features
+
+* new tax api ([#979](https://github.com/medusajs/medusa/issues/979)) ([47588e7](https://github.com/medusajs/medusa/commit/47588e7a8d3b2ae2fed0c1e87fdf1ee2db6bcdc2)), closes [#885](https://github.com/medusajs/medusa/issues/885) [#896](https://github.com/medusajs/medusa/issues/896) [#911](https://github.com/medusajs/medusa/issues/911) [#945](https://github.com/medusajs/medusa/issues/945) [#950](https://github.com/medusajs/medusa/issues/950) [#951](https://github.com/medusajs/medusa/issues/951) [#954](https://github.com/medusajs/medusa/issues/954) [#969](https://github.com/medusajs/medusa/issues/969) [#998](https://github.com/medusajs/medusa/issues/998) [#1017](https://github.com/medusajs/medusa/issues/1017) [#1110](https://github.com/medusajs/medusa/issues/1110)
+
+
+
+
+
+# [0.2.0](https://github.com/medusajs/medusa/compare/medusa-react@0.1.5...medusa-react@0.2.0) (2022-02-25)
+
+
+### Bug Fixes
+
+* use /admin/returns/:id/receive for swap returns ([#1041](https://github.com/medusajs/medusa/issues/1041)) ([7ae754b](https://github.com/medusajs/medusa/commit/7ae754bb6187db17c45b2cfadc625df8f997f5ab))
+
+
+### Features
+
+* new tax api ([#979](https://github.com/medusajs/medusa/issues/979)) ([c56660f](https://github.com/medusajs/medusa/commit/c56660fca9921a3f3637bc137d9794781c5b090f)), closes [#885](https://github.com/medusajs/medusa/issues/885) [#896](https://github.com/medusajs/medusa/issues/896) [#911](https://github.com/medusajs/medusa/issues/911) [#945](https://github.com/medusajs/medusa/issues/945) [#950](https://github.com/medusajs/medusa/issues/950) [#951](https://github.com/medusajs/medusa/issues/951) [#954](https://github.com/medusajs/medusa/issues/954) [#969](https://github.com/medusajs/medusa/issues/969) [#998](https://github.com/medusajs/medusa/issues/998) [#1017](https://github.com/medusajs/medusa/issues/1017) [#1110](https://github.com/medusajs/medusa/issues/1110)
+
+
+
+
+
 ## [0.1.5](https://github.com/medusajs/medusa/compare/medusa-react@0.1.4...medusa-react@0.1.5) (2022-02-06)
 
 ### Bug Fixes